requirements.txt
torchvision>=0.6.0
torch>=1.5.0
Pillow>=7.1.2
setuptools>=46.4.0
numpy>=1.18.4